pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cl-embedded"
version = "0.1.0"
description = "Continual Learning for Resource-Constrained Embedded Systems"
authors = [{ name = "Léonard Rivals" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "torch>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "onnx>=1.14",
    "onnxruntime>=1.16",
    "matplotlib>=3.7",
    "seaborn>=0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "jupyter>=1.0",
    "ipykernel>=6.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
